Flippy_Bit_Bot/requirements.txt
easyocr
pyautogui
Pillow
numpy
pynput
sklearn
matplotlib
pyautogui